This is just a proof of concept, it still needs testing and the bnd.bnd metadata
This creates an interface SemcovConstantsAccessor in the common package.

Two copies of SemconvConstantsAccessorImpl, one in 2.1 and one in 2.0. These provide java methods to get the Semcov constants.

A new copy of MpTelemetryAccessEventMappingUtils in the common package that uess the Accessor; so it is compatible with 2.0 and 2.1 (still to do, there are some differences other than semcov between those versions so both will have to be merged in)

Testing. This is untested so its bound to have some mistakes somewhere